+++ test/actions/navigator.spec.js
@@ -1,6 +1,5 @@
-import expect from 'expect';
-import * as actions from '../../scripts/actions/NavigatorActions';
-import * as types from '../../scripts/constants/ActionTypes';
+import * as actions from '../../client/actions/NavigatorActions';
+import * as types from '../../client/constants/ActionTypes';
 import {mockStore} from '../TestUtils';
 
 describe('navigator actions', () => {
