+++ test/actions/users.spec.js
@@ -27,8 +27,8 @@ jest.mock('global', () => ({
 }));
 
 import { fetch } from 'global';
-import * as actions from '../../scripts/actions/UsersActions';
-import * as types from '../../scripts/constants/ActionTypes';
+import * as actions from '../../client/actions/UsersActions';
+import * as types from '../../client/constants/ActionTypes';
 import { mockStore } from '../TestUtils';
 
 describe('users actions', () => {
